pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-data-analysis-case-library"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "pandas",
  "numpy",
  "scipy",
  "statsmodels",
  "scikit-learn",
  "jupyter",
  "nbconvert",
  "matplotlib",
  "seaborn",
  "linearmodels",
  "lifelines",
  "semopy",
  "pysal",
  "pyyaml",
  "pytest",
]
